@@ -4,7 +4,8 @@
   "description": "Live Code Editor",
   "main": "dist/live-editor.js",
   "scripts": {
-    "build": "webpack --mode development",
+    "lint": "eslint --config ./eslint/eslintrc js/**/*.jsx --fix",
+    "build": "webpack --mode development && gulp build",
     "test": "gulp test"
   },
   "repository": {
@@ -20,15 +21,22 @@
   "devDependencies": {
     "async": "^0.9.0",
     "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.6",
     "babel-loader": "^7.1.5",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-preset-env": "^1.7.0",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "chrome-launcher": "^0.10.2",
     "chrome-remote-interface": "^0.25.2",
     "chrome-unmirror": "^0.1.0",
     "css-loader": "^1.0.0",
-    "eslint": "3.10.2",
+    "eslint": "^4.19.1",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-flowtype": "^2.50.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-react": "^7.10.0",
     "expect.js": "^0.3.1",
     "gulp": "3.9.1",
     "gulp-bower": "0.0.7",
@@ -47,6 +55,7 @@
     "mocha": "^1.21.4",
     "node-static": "^0.7.4",
     "object-assign": "^3.0.0",
+    "prettier": "1.13.7",
     "replace-ext": "0.0.1",
     "request": "^2.44.0",
     "rimraf": "^2.5.4",
@@ -60,7 +69,10 @@
   },
   "dependencies": {
     "@khanacademy/wonder-blocks-button": "^1.0.1",
+    "@khanacademy/wonder-blocks-icon": "^1.0.1",
+    "@khanacademy/wonder-blocks-icon-button": "^1.0.1",
     "@khanacademy/wonder-blocks-modal": "^1.0.1",
+    "@khanacademy/wonder-blocks-progress-spinner": "^1.0.1",
     "ace-builds": "1.3.3",
     "aphrodite": "^2.2.2",
     "backbone": "^1.0.0",
