@@ -1,13 +1,20 @@
 import Button from "@khanacademy/wonder-blocks-button";
 import {OneColumnModal} from "@khanacademy/wonder-blocks-modal";
-import { StyleSheet, css } from 'aphrodite/no-important';
-import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
-import React, {Component} from 'react';
-import 'react-tabs/style/react-tabs.css';
+import {Tab, Tabs, TabList, TabPanel} from "react-tabs";
+import React, {Component} from "react";
+import "react-tabs/style/react-tabs.css";
 
+const i18n = require("i18n");
 const MediaPickerScroller = require("./media-picker-scroller.jsx");
 
 class MediaPickerModal extends Component {
+    props: {
+        mediaClasses: Array<Object>,
+        imagesDir: string,
+        soundsDir: string,
+        onFileSelect: (info: Object) => void,
+        onClose: () => void,
+    };
 
     render() {
         // state: activeClass
@@ -20,34 +27,35 @@ class MediaPickerModal extends Component {
         });
 
         // Now make the tab panels with the media for each class
-        const classesTabPanels = this.props.mediaClasses.map((mediaClass, ind) => {
-            const panelKey = `${mediaClass.className}-panel`;
-            return <TabPanel key={panelKey}>
-                    <MediaPickerScroller
-                        groups={mediaClass.groups}
-                        imagesDir={this.props.imagesDir}
-                        soundsDir={this.props.soundsDir}
-                        onFileSelect={this.props.onFileSelect}
+        const classesTabPanels = this.props.mediaClasses.map(
+            (mediaClass, ind) => {
+                const panelKey = `${mediaClass.className}-panel`;
+                return (
+                    <TabPanel key={panelKey}>
+                        <MediaPickerScroller
+                            groups={mediaClass.groups}
+                            imagesDir={this.props.imagesDir}
+                            soundsDir={this.props.soundsDir}
+                            onFileSelect={this.props.onFileSelect}
                         />
-                </TabPanel>;
-        });
+                    </TabPanel>
+                );
+            },
+        );
 
         return (
             <OneColumnModal
                 content={
                     <Tabs>
-                        <TabList>
-                            {classesTabs}
-                        </TabList>
+                        <TabList>{classesTabs}</TabList>
                         {classesTabPanels}
                     </Tabs>
                 }
                 footer={
-                    <Button onClick={this.props.onClose}>
-                    {i18n._("Ok")}
-                    </Button>
+                    <Button onClick={this.props.onClose}>{i18n._("Ok")}</Button>
                 }
-            />);
+            />
+        );
     }
 }
 
