@@ -1,15 +1,22 @@
 import React, {Component} from "react";
-import Button from "@khanacademy/wonder-blocks-button";
-import { StyleSheet, css } from 'aphrodite/no-important';
+import {StyleSheet, css} from "aphrodite/no-important";
+import {CircularSpinner} from "@khanacademy/wonder-blocks-progress-spinner";
+
+const i18n = require("i18n");
 
 class MediaPickerPreview extends Component {
     // props mediaType, mediaSrc, errorMessage, errorType
+    props: {
+        errorMessage: string,
+        mediaType: string,
+        mediaSrc: string,
+    };
 
     constructor(props) {
         super(props);
         this.state = {
-            imageLoaded: false
-        }
+            imageLoaded: false,
+        };
         this.handleImageLoad = this.handleImageLoad.bind(this);
         this.handleImageError = this.handleImageError.bind(this);
     }
@@ -20,47 +27,52 @@ class MediaPickerPreview extends Component {
 
     handleImageError() {
         this.setState({
-            errorMessage: i18n._("That is not a valid image URL.")
+            errorMessage: i18n._("That is not a valid image URL."),
         });
     }
 
     render() {
         let mediaPreview;
         if (this.props.mediaType === "audio") {
-            mediaPreview = <div>
+            mediaPreview = (
+                <div>
                     <audio
                         className={css(styles.audio)}
                         src={this.props.mediaSrc}
-                        controls/>
-                    <div className={css(styles.error)}></div>
-                </div>;
+                        controls
+                    />
+                    <div className={css(styles.error)} />
+                </div>
+            );
         } else {
             // Only show throbber while we're waiting for image to load
             let loadingImg;
             if (!this.state.imageLoaded && this.props.mediaSrc) {
-                loadingImg = <img
-                    src="/images/spinner.gif"
-                    className={css(styles.spinner)}
-                    />;
+                loadingImg = <CircularSpinner size="small" />;
             }
             let errorDiv;
-            let errorMessage = this.state.errorMessage || this.props.errorMessage;
+            const errorMessage =
+                this.state.errorMessage || this.props.errorMessage;
             if (errorMessage) {
-                errorDiv = <div className={css(styles.error)}>
-                    {errorMessage}
-                    </div>;
+                errorDiv = (
+                    <div className={css(styles.error)}>{errorMessage}</div>
+                );
             }
-            mediaPreview = <div>
+            mediaPreview = (
+                <div>
                     {loadingImg}
                     <div className={css(styles.imgBox)}>
-                        <img className={css(styles.img)}
+                        <img
+                            className={css(styles.img)}
+                            alt={$._("Selected image file")}
                             src={this.props.mediaSrc}
                             onLoad={this.handleImageLoad}
                             onError={this.handleImageError}
                         />
                         {errorDiv}
                     </div>
-                </div>;
+                </div>
+            );
         }
 
         return mediaPreview;
@@ -69,19 +81,19 @@ class MediaPickerPreview extends Component {
 
 const styles = StyleSheet.create({
     audio: {
-        maxWidth: "100%"
+        maxWidth: "100%",
     },
     imgBox: {
         height: "100px",
         width: "100%",
         marginBottom: "5px",
-        textAlign: "center"
+        textAlign: "center",
     },
     img: {
         display: "inline-block",
         maxWidth: "100%",
         maxHeight: "100%",
-        verticalAlign: "middle"
+        verticalAlign: "middle",
     },
     error: {
         display: "inline-block",
@@ -89,14 +101,8 @@ const styles = StyleSheet.create({
         textAlign: "center",
         color: "red",
         fontWeight: "bold",
-        paddingTop: "30%"
+        paddingTop: "30%",
     },
-    spinner: {
-        position: "absolute",
-        top: "14px",
-        left: "50%",
-        marginLeft: "-9px"
-    }
 });
 
 module.exports = MediaPickerPreview;
