@@ -1,4 +1,7 @@
-const $ = require("jquery");
+/* eslint-disable no-console */
+/* TODO: Fix the lint errors */
+/* To fix, remove an entry above, npm run lint, and fix errors. */
+
 const Backbone = require("backbone");
 Backbone.$ = require("jquery");
 
